fixup for "s390/dasd: avoid undefined behaviour"
Martin noticed that the same bug can happen in verify_fcx_max_data.
While at it also remove the whitespace between cast and mdc
to please checkpatch --strict.

please merge.

+++ drivers/s390/block/dasd_eckd.c
@@ -1201,7 +1201,7 @@ static u32 get_fcx_max_data(struct dasd_device *device)
 		dev_warn(&device->cdev->dev, "Detecting the maximum supported data size for zHPF requests failed\n");
 		return 0;
 	} else {
-		return (u32) mdc * FCX_MAX_DATA_FACTOR;
+		return (u32)mdc * FCX_MAX_DATA_FACTOR;
 	}
 }
 
@@ -1220,7 +1220,7 @@ static int verify_fcx_max_data(struct dasd_device *device, __u8 lpm)
 				 mdc, lpm);
 			return mdc;
 		}
-		fcx_max_data = mdc * FCX_MAX_DATA_FACTOR;
+		fcx_max_data = (u32)mdc * FCX_MAX_DATA_FACTOR;
 		if (fcx_max_data < private->fcx_max_data) {
 			dev_warn(&device->cdev->dev,
 				 "The maximum data size for zHPF requests %u "
